backend/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-JWT-Extended>=4.5.0,<5.0.0
Flask-CORS>=4.0.0,<5.0.0
Werkzeug>=3.0.0,<4.0.0
requests>=2.30.0,<3.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.20.0
yfinance>=0.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-flask>=1.0.0
black>=23.0.0
flake8>=6.0.0
